Reward System
This is obviously quite conceptual. I had this idea combined with a bit of stuff I found online, and couldn't figure out how to put it into code, but I think it would be a great addition to BrainBot. The overall idea is that when someone does something nice for someone else, that person could say !thanks to add to the karma supply. To make sure these aren't just given out and are actually meaningful, I think it would be great to have a rather long cool down. Unfortunately I'm not sure how mto implement that, so I'll leave that part up to you :)